[SCB-986] Make InstanceDiscoveryFilter standalone
